+++ app/tag/controller.js
@@ -2,6 +2,13 @@ const Tag = require('./model');
 
 async function store(req, res, next){
     try {
+        let policy = policyFor(req.user);
+        if(!policy.can('create', 'Tag')){
+            return res.json({
+                error: 1, 
+                message: `Anda tidak memiliki akses untuk membuat produk`
+            });
+        }
         params = req.body;
         let tag = new Tag(params);
         await tag.save();
@@ -47,6 +54,13 @@ async function show(req, res, next){
 
 async function destroy(req, res, next){
     try {
+        let policy = policyFor(req.user);
+        if(!policy.can('delete', 'Tag')){
+            return res.json({
+                error: 1, 
+                message: `Anda tidak memiliki akses untuk membuat produk`
+            });
+        }
         let tag = await Tag.findOneAndDelete({
             _id : req.params.id
         });
@@ -61,6 +75,13 @@ async function destroy(req, res, next){
 
 async function update(req, res, next){
     try {
+        let policy = policyFor(req.user);
+        if(!policy.can('update', 'Tag')){
+            return res.json({
+                error: 1, 
+                message: `Anda tidak memiliki akses untuk membuat produk`
+            });
+        }
         params = req.body;
         let tag = await Tag.findOneAndUpdate(
             {_id: req.params.id},
